requirements.txt
fastapi
uvicorn
pillow
pytest
pytest-asyncio
